Добавлен класс NeuralNet. Реализован алгоритм вычисления. Возможны ошибки.

+++ NeuralNet/src/Include/Layer/InputLayer.hpp
@@ -7,6 +7,6 @@ class InputLayer : public Layer<InputNeuron>
 public:
 	InputLayer(size_t n, const std::string& name);
 public:
-	void set_input(const std::vector<int>& values);
+	void set_input(const std::vector<float>& values);
 	void forward();
 };
